# [MLIR][Affine] Fix crash in loop unswitching/hoistAffineIfOp (#130401)