… llvm/test/CodeGen/AMDGPU/materialize-frame-index-sgpr.gfx10.ll
… llvm/test/CodeGen/AMDGPU/materialize-frame-index-sgpr.ll